I'm not sure which I like better.
I do like that the `make` version parses out, for the eyes, more
quickly what you're working with.

But this form's more typical. And, for particular types (like
`gtk-menu-item-new-with-mnemonic`), it's easy to figure out how to
create and there's easy, documentation that can be looked up by the
function name; of course, Geiser's completely broken for me with this
library so, perhaps, there would be a way to figure out the `make`
equivalent. If I come across any documentation that explains, I'll
make note of it.