Good Hell, this one was a pain to figure out…
Mostly it was the enums/constants; I put comments for where I finally
figured out what was supposed to be what but the fact that these
aren't already defined, in some way (or maybe they are; but Hell if I
could find the documentation!), is disheartening.

I'm counting on the keyval to sort of be able to be pieced back
together from the function that's used (it just amounts to the number
113) but I was able to get there from these Ruby doc.s:
https://ruby-gnome2.osdn.jp/de/hiki.cgi?Gdk%3A%3AKeyval#GDK_Q.

Also, the documentation says you can pass `#f` for
`gtk-image-menu-item-new-from-stock` to not associate an `accel-group`
but that didn't work so we improvised.

Otherwise, about as straightforward as the previous examples.